requirements.txt
Flask
google-auth
google-auth-oauthlib
google-auth-httplib2
requests
cachecontrol
fints
firebase_admin
pycryptodome